+++ CHANGELOG.md
@@ -6,6 +6,7 @@ MumbleDJ Changelog
 * Removed `sanitize` dependency.
 * Reworked `Makefile` slightly.
 * Now uses `gumbleutil.PlainText` for removing HTML tags instead of `sanitize`.
+* Added `!setcomment` which allows admin users to set the comment for the bot.
 
 ### February 3, 2015 -- `v2.4.1`
 * Made it possible to place MumbleDJ binary in `~/bin` instead of `/usr/local/bin` if the folder exists.